extract token span utilities into token-merge core crate
