+++ Cargo.toml
@@ -25,6 +25,7 @@ members = [
   "crates/bitnet-inference-metrics-core", # SRP: shared inference timing/throughput metric DTOs
   "crates/bitnet-rope",
   "crates/bitnet-tokenizers",
+  "crates/bitnet-token-merge-core", # SRP: reusable token span merge/split primitives
   "crates/bitnet-tokenizer-model-core", # SRP: tokenizer model/vocabulary detection helpers
   "crates/bitnet-tokenizer-discovery-core", # SRP: tokenizer path auto-discovery core
   "crates/bitnet-prompt-templates",
@@ -147,6 +148,7 @@ default-members = [
   "crates/bitnet-test-fixtures-core",
   "crates/bitnet-models",
   "crates/bitnet-tokenizers",
+  "crates/bitnet-token-merge-core", # SRP: reusable token span merge/split primitives
   "crates/bitnet-tokenizer-model-core",
   "crates/bitnet-tokenizer-discovery-core",
   "crates/bitnet-cpu-activations", # SRP: CPU activation kernels shared microcrate
