+++ crates/bitnet-token-merge-core/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "bitnet-token-merge-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "SRP core token span merge/split utilities"
+
+[dependencies]
+
+[dev-dependencies]
+proptest = { workspace = true }
